backend_py/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
httpx>=0.27.0
sentence-transformers>=3.0.0
google-genai>=0.3.0
elevenlabs>=1.0.0
structlog>=24.0.0
slowapi>=0.1.9